requirements.txt
flask
pandas
scikit-learn
rdkit-pypi
joblib
python-dotenv
matplotlib
seaborn
mysql-connector-python
nltk